requirements.txt
colorlog==6.7.0
mparticle==0.15.0
setuptools==65.5.1
snowflake_connector_python[pandas]==3.0.0
urllib3==1.26.14
snowflake-snowpark-python==1.1.0
Jinja2==3.1.2
boto3==1.26.75
botocore==1.29.77
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
Faker==18.4.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
yarl==1.8.2